docs: Add code coverage for GSON
As a part of the task it is good to have a code coverage report. This
will tell us which parts of the code does not have enough tests.
This will also make the choice of 5 methods to make new tests for
easier, since we can choose those that do not have sufficient tests (see
issue #2 and #3)