feat: Implement branch coverage classpath
To make issue #1 easier, it is good to have a class that automatically makes our branch coverage DIY.
The code added in this commit works, and does print which branches are not properly tested.
However, these branches are only referred to with their branch ID's (which are sudo random).
Therefore, this has to be solved before we can close the realted issue.