feat: Implement automatic branch coverage
The problems that were talked about in the previous commit have now been finished.
It is also possible to see percentage of missed branches that this results in.

Since this is automatic it will now close issue #1. Furtherwork, on this tool may be needed. However, that will require a new issue.

Resolves #1